tests/requirements.txt
# RiftLink tests
pytest>=7.0.0
pyserial>=3.5
